+++ lib/src/sp3d_v2d.dart
@@ -13,7 +13,7 @@ import 'package:flutter/material.dart';
 @immutable
 class Sp3dV2D {
   final String className = 'Sp3dV2D';
-  final String version = '5';
+  final String version = '6';
   final double x;
   final double y;
 
@@ -22,6 +22,7 @@ class Sp3dV2D {
   /// * [y] : y.
   const Sp3dV2D(this.x, this.y);
 
+  /// Deep copy this object.
   Sp3dV2D deepCopy() {
     return Sp3dV2D(x, y);
   }
@@ -33,6 +34,7 @@ class Sp3dV2D {
     return Sp3dV2D(x ?? this.x, y ?? this.y);
   }
 
+  /// Convert to Map.
   Map<String, dynamic> toDict() {
     Map<String, dynamic> d = {};
     d['class_name'] = className;
@@ -42,6 +44,7 @@ class Sp3dV2D {
     return d;
   }
 
+  /// convert from Map.
   static Sp3dV2D fromDict(Map<String, dynamic> src) {
     return Sp3dV2D(src['x'], src['y']);
   }
@@ -82,6 +85,7 @@ class Sp3dV2D {
   /// (en)Compare while considering the error. Returns true if x, y are all within the e_range.
   ///
   /// (ja)誤差を考慮しつつ比較します。x, y の全てが誤差e_range以内の場合はtrueを返します。
+  ///
   /// * [other] : other vector.
   /// * [eRange] : The range of error to allow. This must be a positive number.
   bool equals(Sp3dV2D other, double eRange) {
@@ -89,6 +93,19 @@ class Sp3dV2D {
         other.x <= x + eRange &&
         y - eRange <= other.y &&
         other.y <= y + eRange;
+  }
+
+  /// (en)Returns a new vector that rotates this vector.
+  ///
+  /// (ja)このベクトルを回転した新しいベクトルを返します。
+  ///
+  /// * [origin] : the origin of rotation.
+  /// * [radian] : radian = degree * pi / 180.
+  Sp3dV2D rotated(Sp3dV2D origin, double radian) {
+    final Sp3dV2D diff = this - origin;
+    final double r = diff.len();
+    final double theta = atan2(diff.y, diff.x) + radian;
+    return Sp3dV2D(origin.x + r * cos(theta), origin.y + r * sin(theta));
   }
 
   Sp3dV2D operator +(Sp3dV2D v) {